Added turn count and Item buttons
We added a turn count that increases each time a button is pressed. We
changed the clamp so that it is an if statement in the combat class. I
created blank classes for every item. We corrected the stats for fists
and pistol. We created the first aid kit item. We created 2 more buttons
medical ablity and assist ability and linked all the abilities to the
classes to call them in the combat class