Overhauled the combat class entirely
Made it so that the abilities are implemented inside of their respective
classes rather than in the Combat class itself.
Also rewrote and cleaned up GlobalVariables, now relying on that much
less. Also created an Entity super class for the Enemy object and Player
object. Sepreated the Player into PlayerHandler which manages movement
in scene1, and Player, which manages the player's inventory and other
combat realted stuff.