Improved Combat Class with Timing
Added a method wait that adds a 2 second wait time between the players
attack and the enemies attack. Added the methods updatePlayer and
updateEnemy in combat to update their health and damage done in the
method. Also added eHealed and pHealed varaiables which displays a
positive value in green for hp added similar in how damage done displays
hp lost in red. And Jackson is gay. Also added the method enemyAttack in
combat so that the enemy can attack at a time after the player attacks.
Disabled the buttons after they are pressed while they are waiting.